[FrontEnd] 파일 구조 수정, edit mode 전환 작업중
