+++ src/pages/editor.tsx
@@ -1,10 +1,10 @@
 import React, { LegacyRef } from 'react'
 import 'codemirror/lib/codemirror.css'
 import '@toast-ui/editor/dist/toastui-editor.css'
-
 import { Editor } from '@toast-ui/react-editor'
+import { Button, Row, Col } from 'antd';
 
-function WritePage() {
+export default function MemoEditor() {
   const editorRef: LegacyRef<Editor> = React.createRef()
 
   const handleClick = () => {
@@ -13,17 +13,22 @@ function WritePage() {
 
   return (
     <>
+      <Row justify="end">
+        <Col span={1}>
+          <Button danger>취소</Button>
+        </Col>
+        <Col span={1}>
+          <Button type="primary" onClick={handleClick}>저장</Button>
+        </Col>
+      </Row>
       <Editor
         previewStyle="vertical"
         height="400px"
         initialEditType="wysiwyg"
         // TODO: 수정할 때 initial에 기존 메모 값 가져오기
-        initialValue="메모를 입력하세요"
+        initialValue={"메모를 입력하세요"}
         ref={editorRef}
       />
-      <button onClick={handleClick}>submit</button>
     </>
   )
 }
-
-export default WritePage
